pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pocket-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "colorlog>=6.10.1",
    "fastmcp>=2.13.0.2",
    "openai>=2.6.1",
    "python-dotenv>=1.2.1",
    "semantic-kernel>=1.36.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
